# Improve performance for initialize