`file:is-match`:

**as-is**
total 0.356sec (50,000calls, 7microsec/call)

**to-be**
total 0.005sec (50,000calls, 0microsec/call) 🆙 

`ast-index/base:create-indexes`:

**as-is**
total 1.169sec (2calls)

**to-be**
total 0.25sec (2calls) 🆙